requirement.txt
asgiref==3.6.0
attrs==22.2.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.5
django-cors-headers==3.13.0
django-environ==0.9.0
django-phonenumber-field==7.0.2
djangorestframework==3.14.0
drf-yasg==1.21.4
idna==3.4
importlib-resources==5.10.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.1
packaging==23.0
phonenumbers==8.13.4
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
pyrsistent==0.19.3
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sqlparse==0.4.3
swagger-spec-validator==3.0.3
typing_extensions==4.4.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
zipp==3.11.0
